Lista 2

+++ lista02/2.c
@@ -1,18 +1,18 @@
 #include <stdio.h>
 
 
-/* Função recursiva para verificar se um número é par ou ímpar */
-double V(double l, double c, double h)
-{	
+/* Função para encontrar o volume de um paralelepípedo */
+float V(float l, float c, float h)
+{
 	return (l*c*h);
 }
 
 
 int main()
 {
-	double l, c, h;
-	scanf("%lf %lf %lf", &l, &c, &h);
-	printf("%lf\n", V(l, c, h));
+	float l, c, h;
+	scanf("%f %f %f", &l, &c, &h);
+	printf("%.2f\n", V(l, c, h));
 
 	return 0;
 }
